requirements.txt
    langchain-core>=0.3.55
    langchain-openai>=0.3.14
    langchain-groq>=0.1.3
    langgraph>=0.2.76
    pandas>=2.2.3
    plotly>=6.0.1
    psycopg2-binary>=2.9.10
    pydantic>=2.11.3
    python-dotenv>=1.1.0
    sqlalchemy>=2.0.40
    tabulate>=0.9.0
    kaleido>=0.2.1
    fastapi>=0.104.1
    uvicorn>=0.24.0